requirements.txt
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.12.0
python-dotenv==1.0.0
numpy==1.21.0
Flask==2.1.0
Flask-CORS==3.0.10
APScheduler==3.9.1